+++ src/index/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_subdirectory(vocabulary)
 add_library(index
-        Index.h Index.cpp Index.Text.cpp
+        Index.h Index.cpp IndexImpl.Text.cpp
         Vocabulary.h Vocabulary.cpp
         VocabularyGenerator.h VocabularyGeneratorImpl.h
         ConstantsIndexBuilding.h
